atepp cleanup
